requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Environment and Configuration
python-dotenv==1.0.0

# Networking
requests==2.31.0
websockets==12.0
aiohttp==3.9.1

# API Integrations
twilio==8.10.3
openai==1.3.7
deepgram-sdk==3.2.7
elevenlabs==0.2.26
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Event Handling
event-emitter==0.9.0

# Logging
colorlog==6.8.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# AI and ML
anthropic==0.7.7
numpy==1.25.2

# RAG Dependencies
qdrant-client>=1.7.0
sentence-transformers>=2.2.0

# Additional utilities
pydantic==2.5.0
python-multipart==0.0.6
